requirements.txt
-i https://pypi.org/simple
alembic==1.13.0; python_version >= '3.8'
aniso8601==9.0.1
appnope==0.1.3; sys_platform == 'darwin'
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.1.1; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.7'
executing==2.0.1; python_version >= '3.5'
faker==20.1.0
flask==2.2.5
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-marshmallow==0.15.0
flask-migrate==4.0.5
flask-restful==0.3.10
flask-sqlalchemy==3.0.3
idna==3.6; python_version >= '3.5'
importlib-metadata==7.0.0; python_version < '3.10'
importlib-resources==6.1.1; python_version < '3.9'
ipdb==0.13.9
ipython==8.12.3; python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
joblib==1.3.2; python_version >= '3.7'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
marshmallow==3.20.1; python_version >= '3.8'
marshmallow-sqlalchemy==0.29.0
matplotlib-inline==0.1.6; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
parso==0.8.3; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.41; python_full_version >= '3.7.0'
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
setuptools==69.0.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.23; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
stack-data==0.6.3
stripe==7.8.2
threadpoolctl==3.2.0; python_version >= '3.8'
toml==0.10.2; python_version >= '3.7'
tqdm==4.66.1; python_version >= '3.7'
traitlets==5.14.0; python_version >= '3.8'
typing-extensions==4.9.0; python_version < '3.9'
urllib3==2.1.0; python_version >= '3.8'
wcwidth==0.2.12
werkzeug==2.2.2
zipp==3.17.0; python_version >= '3.8'
